// models/type.model.ts
export enum PokemonType {
  NORMAL = "Normal",
  FIRE = "Fire",
  WATER = "Water",
  GRASS = "Grass",
  FLYING = "Flying",
  FIGHTING = "Fighting",
  POISON = "Poison",
  ELECTRIC = "Electric",
  GROUND = "Ground",
  ROCK = "Rock",
  PSYCHI = "Psychic",
  ICE = "Ice",
  BUG = "Bug",
  GHOST = "Ghost",
  STEEL = "Steel",
  DRAGON = "Dragon",
  DARK = "Dark",
  FAIRY = "Fairy",
}

export enum PokemonTypeColor {
  NORMAL = "bg-green-500",
  FIRE = "bg-red-600",
  WATER = "bg-blue-200",
  GRASS = "bg-green-400",
  FLYING = "bg-blue-100",
  FIGHTING = "bg-red-700",
  POISON = "bg-pink-500",
  ELECTRIC = "bg-yellow-300",
  GROUND = "bg-yellow-900",
  ROCK = "bg-yellow-700",
  PSYCHI = "bg-purple-400",
  ICE = "bg-blue-50",
  BUG = "bg-green-200",
  GHOST = "bg-purple-50",
  STEEL = "bg-red-200",
  DRAGON = "bg-purple-700",
  DARK = "bg-blue-900",
  FAIRY = "bg-pink-400",
}

export interface TypeBadgeProps {
  type: string;
}

export enum BACKGROUND_COLOR {
  GREEN = "bg-green-100 text-gray-700",
  BLUE = "bg-blue-100 text-gray-700",
  BROWN = "bg-yellow-100 text-gray-700",
  PINK = "bg-pink-100 text-gray-700",
  YELLOW = "bg-yellow-100 text-gray-700",
  RED = "bg-red-100 text-gray-700",
  PURPLE = "bg-purple-100 text-gray-700",
  GRAY = "bg-gray-100 text-gray-700",
}

export enum BORDER_COLOR {
  GREEN = "border-green-300 text-white",
  BLUE = "border-blue-400 text-white",
  BROWN = "border-yellow-600 text-white",
  PINK = "border-pink-400 text-white",
  YELLOW = "border-yellow-400 text-white",
  RED = "border-red-600 text-white",
  PURPLE = "border-purple-400 text-white",
  GRAY = "border-gray-500 text-white",
}

export enum TYPE_VARIANTS {
  GREEN = "bg-green-400 text-white",
  BLUE = "bg-blue-400 text-white",
  BROWN = "bg-yellow-900 text-white",
  PINK = "bg-pink-400 text-white",
  YELLOW = "bg-yellow-500 text-white",
  RED = "bg-red-600 text-white",
  PURPLE = "bg-purple-400 text-white",
  GRAY = "bg-gray-500 text-white",
}